+++ src/typescript-worker/build.ts
@@ -4,28 +4,25 @@
  * SPDX-License-Identifier: BSD-3-Clause
  */
 
-import {TypeScriptBuilder} from './typescript.js';
 import {BareModuleTransformer} from './bare-module-transformer.js';
 import {ImportMapResolver} from './import-map-resolver.js';
 
 import type {
   SampleFile,
-  ModuleImportMap,
   BuildOutput,
+  TypeScriptWorkerConfig,
 } from '../shared/worker-api.js';
 import {CachingCdn} from './caching-cdn.js';
+import {getBuilder} from './project-cache.js';
 
 export const build = async (
   files: Array<SampleFile>,
-  config: {
-    importMap: ModuleImportMap;
-    cdnBaseUrl?: string;
-  },
+  config: TypeScriptWorkerConfig,
   emit: (result: BuildOutput) => void
 ): Promise<void> => {
   const moduleResolver = new ImportMapResolver(config.importMap);
   const cdn = new CachingCdn(config.cdnBaseUrl ?? 'https://unpkg.com/');
-  const tsBuilder = new TypeScriptBuilder(cdn, moduleResolver);
+  const tsBuilder = getBuilder(config, cdn, moduleResolver);
   const bareModuleBuilder = new BareModuleTransformer(cdn, moduleResolver);
   const results = bareModuleBuilder.process(
     tsBuilder.process(files.map((file) => ({kind: 'file', file})))