+++ src/typescript-worker/typescript.ts
@@ -37,6 +37,16 @@ export class TypeScriptBuilder {
   private readonly _cdn: CachingCdn;
   private readonly _importMapResolver: ImportMapResolver;
 
+  private readonly _languageServiceHost = new WorkerLanguageServiceHost(
+    self.origin,
+    compilerOptions
+  );
+
+  private readonly _languageService = ts.createLanguageService(
+    this._languageServiceHost,
+    ts.createDocumentRegistry()
+  );
+
   constructor(cdn: CachingCdn, importMapResolver: ImportMapResolver) {
     this._cdn = cdn;
     this._importMapResolver = importMapResolver;
@@ -79,23 +89,21 @@ export class TypeScriptBuilder {
       loadedFiles.set(url, file.content);
     }
 
-    // Fast initial compile for JS emit and syntax errors.
-    const languageServiceHost = new WorkerLanguageServiceHost(
-      loadedFiles,
-      self.origin,
-      compilerOptions
-    );
-    const languageService = ts.createLanguageService(
-      languageServiceHost,
-      ts.createDocumentRegistry()
-    );
-    const program = languageService.getProgram();
+    // Sync the new loaded files with the servicehost.
+    // If the file is missing, it's added, if the file is modified,
+    // the modification data and versioning will be handled by the servicehost.
+    // If a file is removed, it will be removed from the file list
+    this._languageServiceHost.sync(loadedFiles);
+
+    const program = this._languageService.getProgram();
     if (program === undefined) {
       throw new Error('Unexpected error: program was undefined');
     }
 
     for (const {file, url} of inputFiles) {
-      for (const tsDiagnostic of languageService.getSyntacticDiagnostics(url)) {
+      for (const tsDiagnostic of this._languageService.getSyntacticDiagnostics(
+        url
+      )) {
         yield {
           kind: 'diagnostic',
           filename: file.name,
@@ -129,12 +137,12 @@ export class TypeScriptBuilder {
       // TypeScript is going to look for these files as paths relative to our
       // source files, so we need to add them to our filesystem with those URLs.
       const url = new URL(`node_modules/${path}`, self.origin).href;
-      if (!loadedFiles.has(url)) {
-        loadedFiles.set(url, content);
-      }
+      this._languageServiceHost.updateFileContentIfNeeded(url, content);
     }
     for (const {file, url} of inputFiles) {
-      for (const tsDiagnostic of languageService.getSemanticDiagnostics(url)) {
+      for (const tsDiagnostic of this._languageService.getSemanticDiagnostics(
+        url
+      )) {
         yield {
           kind: 'diagnostic',
           filename: file.name,
@@ -145,19 +153,59 @@ export class TypeScriptBuilder {
   }
 }
 
+interface VersionedFile {
+  version: number;
+  content: string;
+}
+
 class WorkerLanguageServiceHost implements ts.LanguageServiceHost {
   readonly compilerOptions: ts.CompilerOptions;
   readonly packageRoot: string;
-  readonly files: Map<string, string>;
+  readonly files: Map<string, VersionedFile> = new Map<string, VersionedFile>();
 
-  constructor(
-    files: Map<string, string>,
-    packageRoot: string,
-    compilerOptions: ts.CompilerOptions
-  ) {
+  constructor(packageRoot: string, compilerOptions: ts.CompilerOptions) {
     this.packageRoot = packageRoot;
     this.compilerOptions = compilerOptions;
-    this.files = files;
+  }
+
+  /*
+   *  When a new new "process" command is received, we iterate through all of the files,
+   *  and update files accordingly depending on if they have new content or not.
+   *
+   *  With how the TS API works, we can use simple versioning to tell the
+   *  Language service that a file has been updated
+   *
+   *  If the file submitted is a new file, we add it to our collection
+   */
+  updateFileContentIfNeeded(fileName: string, content: string) {
+    const file = this.files.get(fileName);
+    if (file && file.content !== content) {
+      file.content = content;
+      file.version += 1;
+    } else {
+      this.files.set(fileName, {content, version: 0});
+    }
+  }
+
+  /**
+   * Sync up the freshly acquired project files.
+   * In the syncing process files yet to be added are added, and versioned.
+   * Files that existed already but are modified are updated, and their version number
+   * gets bumped fo that the languageservice knows to update these files.
+   * */
+  sync(files: Map<string, string>) {
+    files.forEach((file, fileName) =>
+      this.updateFileContentIfNeeded(fileName, file)
+    );
+    this._removeDeletedFiles(files);
+  }
+
+  private _removeDeletedFiles(files: Map<string, string>) {
+    this.getScriptFileNames().forEach((fileName) => {
+      if (!files.has(fileName)) {
+        this.files.delete(fileName);
+      }
+    });
   }
 
   getCompilationSettings(): ts.CompilerOptions {
@@ -168,8 +216,8 @@ class WorkerLanguageServiceHost implements ts.LanguageServiceHost {
     return [...this.files.keys()];
   }
 
-  getScriptVersion() {
-    return '-1';
+  getScriptVersion(fileName: string) {
+    return this.files.get(fileName)?.version.toString() ?? '-1';
   }
 
   fileExists(fileName: string): boolean {
@@ -177,7 +225,7 @@ class WorkerLanguageServiceHost implements ts.LanguageServiceHost {
   }
 
   readFile(fileName: string): string | undefined {
-    return this.files.get(fileName);
+    return this.files.get(fileName)?.content;
   }
 
   getScriptSnapshot(fileName: string): ts.IScriptSnapshot | undefined {